Changed the event forwarding policy to be whitelist-only
Now only certain backbone events will be forwarded. In addition I
added a check for 'change' and 'change:[attribute]' events, which
will only be forwarded if the model is still in the collection.

This strange condition was occuring with the
backbone-filtered-collection module, where a model would change to
not fit a filter function, would be removed from the collection,
triggering an event, but then the 'change' event would get forwarded
for a model that was no longer in the collection.